pyproject.toml
[project]
name = "cendry"
version = "0.0.1"
description = "A Firestore ODM for Python"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-firestore>=2.19",
    "anyio>=4",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-bdd>=8",
    "anyio>=4",
    "mypy>=1",
    "ty",
    "coverage>=7",
    "ruff>=0.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cendry"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
strict = true
packages = ["cendry"]
mypy_path = "src"

[tool.coverage.run]
source = ["cendry"]

[tool.coverage.report]
fail_under = 100
show_missing = true

[tool.ruff]
target-version = "py313"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "F",      # pyflakes
    "E", "W", # pycodestyle
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # ruff-specific
    "PT",     # flake8-pytest-style
    "PIE",    # flake8-pie
    "T20",    # flake8-print
    "PYI",    # flake8-pyi
    "FURB",   # refurb
    "PERF",   # perflint
    "FLY",    # flynt
    "FA",     # flake8-future-annotations
    "ISC",    # flake8-implicit-str-concat
    "RET",    # flake8-return
    "TID",    # flake8-tidy-imports
]
ignore = [
    "UP040",  # TypeAlias — we use type statements instead when applicable
]

[tool.ruff.lint.isort]
known-first-party = ["cendry"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]
